add: feedback reply container component for reply users
